+++ Assets/Scripts/Pedestrian/PedestrianPosition.cs
@@ -12,10 +12,10 @@ using System;
 public class PedestrianPosition
 {
 	private int id;
-	private decimal time;
+	private float time;
 	private float x;
 	private float y;
-	public PedestrianPosition (int id, decimal time, float x, float y) {
+	public PedestrianPosition (int id, float time, float x, float y) {
 		this.id = id;
 		this.time = time;
 		this.x = x;
@@ -23,7 +23,7 @@ public class PedestrianPosition
 	}
 
 	public int getID() {return this.id;}
-	public decimal getTime() {return this.time;}
+	public float getTime() {return this.time;}
 	public float getX() {return this.x;}
 	public float getY() {return this.y;}
 }
